school_test_score_management/src/main/java/org/codesquad/util/AutoIDUtil.java
package org.codesquad.util;

public class AutoIDUtil {

    private static int studentId = 1001;
    private static int subjectId = 1001;

    public static String createStudentId() {
        return String.valueOf(studentId++);
    }

    public static String createStudentId(String id) {
        int number = Integer.parseInt(id);

        if (studentId < number) {
            studentId = number;
        }

        return String.valueOf(studentId++);
    }

    public static String createSubjectId() {
        return "subject" + subjectId++;
    }
}
